backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
duckdb>=1.0.0
pandas>=2.0.0
google-generativeai>=0.8.0
langchain-text-splitters>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
